[ESIMD] Fix stencil tests timeouts
I analyzed the history of ESIMD stencil tests in CI and found that
stencil tests are sensitive to the time limit. I.e. they are not
hanging but timing out before finishing the job.
Since #148 Stencil.cpp was failing several times per day.

This patch reduces back the working set for the two tests allowing
them to finish before in required time limit. Also, added a test
parameter to run with a bigger dimension size.