[SYCL][CUDA] Corrected a missing triple and set an XFAIL: cuda.
Now that the fp16 aspect is connected to the cuda PI (https://github.com/intel/llvm/pull/4029/files) one test case (fp16-with-unnamed-lambda.cpp) that now runs for cuda if the device has the fp16 aspect failed because it was missing the triple for ptx.  The triple has been added.
nan.cpp fails for the fp64 case that was switched on by (https://github.com/intel/llvm/pull/3950).

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>