Adjust enqueue-arg-order-image test
With a recent update to the runtime pull request blocked commands are no
longer waited for during enqueueing of their users. This introduces more
cases where host tasks are used to "glue" events from different
contexts, which leads to a change to expected order of PI calls in the
test.